secure_platform/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiosqlite>=0.19.0
cryptography>=41.0.0
numpy>=1.26.0
networkx>=3.0
